Refactor Roo::Excelx to use new formatting system
The Roo::Excelx class was refactored to use the new formatting system. This simplifies the logic needed to correctly determine the format of a cell and also facilitates true untyped data return.  For untyped data return, the data is returned in a string format that ActiveRecord can appropriately parse into Ruby data types.